+++ src/components/Concentrator/ConcentratorVaultDepositForm.tsx
@@ -20,8 +20,8 @@ export const ConcentratorVaultDepositForm: FC<VaultProps> = (props) => {
   return (
     <VaultDepositForm
       {...props}
-      defaultInputToken={props.vaultAddress}
-      defaultOutputToken={concentrator?.data?.ybTokenAddress}
+      initInputToken={props.vaultAddress}
+      initOutputToken={concentrator?.data?.ybTokenAddress}
       underlyingAssets={underlyingAssets}
       productType="concentrator"
     />
