+++ src/components/Concentrator/ConcentratorVaultWithdrawForm.tsx
@@ -20,8 +20,8 @@ export const ConcentratorVaultWithdrawForm: FC<VaultProps> = (props) => {
   return (
     <VaultWithdrawForm
       {...props}
-      defaultInputToken={concentrator?.data?.ybTokenAddress}
-      defaultOutputToken={props.vaultAddress}
+      initInputToken={concentrator?.data?.ybTokenAddress}
+      initOutputToken={props.vaultAddress}
       underlyingAssets={underlyingAssets}
       productType="concentrator"
     />
