+++ src/hooks/lib/preview/compounder/useTokenPreviewRedeemFallback.ts
@@ -1,10 +1,14 @@
+import { useEffect, useState } from "react"
 import { Address, useContractRead } from "wagmi"
 
-import useTokenPreviewRedeemUnderlying from "@/hooks/lib/preview/compounder/useTokenPreviewRedeemUnderlying"
+import { getGlpPrice } from "@/lib/api/pricer/getGlpPrice"
 import { useActiveChainId } from "@/hooks/useActiveChainId"
 import { useTokenVaultSymbol } from "@/hooks/useTokenVaultSymbol"
 
-import { AMMCompounderBase } from "@/constant/abi"
+import { AMMCompounderBase, GlpVault } from "@/constant/abi"
+import { ETH, glpVault } from "@/constant/addresses"
+
+const WETH_ARBI: Address = "0x82aF49447D8a07e3bd95BD0d56f35241523fBab1"
 
 export default function useTokenPreviewRedeemFallback({
   asset,
@@ -52,21 +56,87 @@ export default function useTokenPreviewRedeemFallback({
     enabled: isUnderlyingAsset && enabled,
   })
 
-  if (isUnderlyingAsset && ybTokenSymbol === "fcGLP") {
-    return {
-      ...previewUnderlying,
-      data: {
-        minAmountWei: BigInt(previewUnderlying?.data ?? 0).toString(),
-        resultWei: BigInt(previewUnderlying?.data ?? 0).toString(),
-      },
-    }
+  return isUnderlyingAsset && ybTokenSymbol === "fcGLP"
+    ? {
+        ...previewUnderlying,
+        data: {
+          minAmountWei: BigInt(previewUnderlying?.data ?? 0).toString(),
+          resultWei: BigInt(previewUnderlying?.data ?? 0).toString(),
+        },
+      }
+    : {
+        ...preview,
+        data: {
+          minAmountWei: undefined,
+          resultWei: (preview.data ?? BigInt(0)).toString(),
+        },
+      }
+}
+
+function useTokenPreviewRedeemUnderlying({
+  token,
+  amount,
+  slippage,
+  enabled,
+}: {
+  token?: Address
+  amount: string
+  slippage: number
+  enabled: boolean
+}) {
+  const chainId = useActiveChainId()
+  const isArbitrumFamily = chainId === 313371 || chainId === 42161
+  const [glpPrice, setGlpPrice] = useState(0)
+
+  token = token ?? "0x"
+  if (token == ETH) {
+    token = WETH_ARBI
   }
 
+  useEffect(() => {
+    getGlpPrice().then((price) => setGlpPrice(price))
+  }, [])
+
+  const usdgAmount = Number(amount) * glpPrice
+  const redemptionAmount = useContractRead({
+    chainId,
+    abi: GlpVault,
+    address: glpVault,
+    enabled: isArbitrumFamily && enabled,
+    functionName: "getRedemptionAmount",
+    args: [token, BigInt(usdgAmount)],
+  })
+
+  const MINT_BURN_FEE_BASIS_POINTS = 25
+  const TAX_BASIS_POINTS = 50
+  const BASIS_POINTS_DIVISOR = 10000
+
+  const feeBasisPoints = useContractRead({
+    chainId,
+    abi: GlpVault,
+    address: glpVault,
+    enabled: isArbitrumFamily && enabled,
+    functionName: "getFeeBasisPoints",
+    args: [
+      token,
+      BigInt(usdgAmount),
+      BigInt(MINT_BURN_FEE_BASIS_POINTS),
+      BigInt(TAX_BASIS_POINTS),
+      false,
+    ],
+  })
+
+  const amountOut = !isArbitrumFamily
+    ? 0
+    : Number(redemptionAmount.data ?? 0) *
+      ((BASIS_POINTS_DIVISOR - Number(feeBasisPoints.data ?? 0)) /
+        BASIS_POINTS_DIVISOR)
+
+  const amountOutWithSlippage = Number(amountOut) * (1 + slippage)
   return {
-    ...preview,
-    data: {
-      minAmountWei: undefined,
-      resultWei: (preview.data ?? BigInt(0)).toString(),
-    },
+    isLoading: false,
+    isFetching: false,
+    isSuccess: true,
+    data: parseInt(amountOutWithSlippage.toFixed(2)),
   }
 }
