--- src/hooks/lib/preview/compounder/useTokenPreviewRedeemUnderlying.ts
@@ -1,22 +1,0 @@
-import { Address } from "wagmi"
-
-import useTokenPreviewRedeem from "@/hooks/lib/preview/compounder/useTokenPreviewRedeem"
-
-export default function useTokenPreviewRedeemUnderlying({
-  token,
-  amount,
-  slippage,
-  enabled,
-}: {
-  token?: Address
-  amount: string
-  slippage: number
-  enabled: boolean
-}) {
-  return useTokenPreviewRedeem({
-    token,
-    amount,
-    slippage,
-    enabled,
-  })
-}
