+++ src/hooks/lib/preview/useCurvePreviewDepositUnderlying.ts
@@ -1,7 +1,6 @@
 import { Address, useContractRead, useContractReads } from "wagmi"
 
 import { VaultType } from "@/lib/types"
-import useCalcTokenAmount from "@/hooks/lib/preview/compounder/useCalcTokenAmount"
 import { useVaultContract } from "@/hooks/lib/useVaultContract"
 import { useActiveChainId } from "@/hooks/useActiveChainId"
 
@@ -12,9 +11,10 @@ import {
   crvTwoCryptoTokenAddress,
   ETH,
   fraxBpTokenAddress,
-  WETH,
-  WETH_ARBI,
 } from "@/constant/addresses"
+
+const WETH_ARBI: Address = "0x82aF49447D8a07e3bd95BD0d56f35241523fBab1"
+const WETH: Address = "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2"
 
 export default function useCurvePreviewDeposit({
   asset,
@@ -108,4 +108,31 @@ export default function useCurvePreviewDeposit({
     },
   }
   //}
+}
+
+function useCalcTokenAmount({
+  asset,
+  assets,
+  isDeposit,
+  enabled,
+}: {
+  asset: Address
+  assets: bigint[]
+  isDeposit: boolean
+  enabled: boolean
+}) {
+  const chainId = useActiveChainId()
+  const poolCurveAddress = ARBI_CURVE_ADDRESS[asset] ?? "0x"
+
+  return useContractRead({
+    chainId,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    abi: (poolCurveAddress === crvTriCryptoPoolAddress
+      ? CurvePool3Assets
+      : CurvePool2Assets) as any,
+    address: poolCurveAddress,
+    enabled: !!asset && enabled,
+    functionName: "calc_token_amount",
+    args: [assets, isDeposit],
+  })
 }
