+++ src/hooks/lib/preview/useCurvePreviewRedeemUnderlying.ts
@@ -1,19 +1,17 @@
-import { Address, useContractReads } from "wagmi"
+import { Address, useContractRead, useContractReads } from "wagmi"
 
 import { VaultType } from "@/lib/types"
-import useCalcWithdrawOneCoin from "@/hooks/lib/preview/compounder/useCalcWithdrawOneCoin"
 import { useActiveChainId } from "@/hooks/useActiveChainId"
 
 import { CurvePool2Assets, CurvePool3Assets } from "@/constant/abi"
 import {
   ARBI_CURVE_ADDRESS,
   crvTriCryptoPoolAddress,
-  crvTwoCryptoTokenAddress,
   ETH,
-  fraxBpTokenAddress,
-  WETH,
-  WETH_ARBI,
 } from "@/constant/addresses"
+
+const WETH_ARBI: Address = "0x82aF49447D8a07e3bd95BD0d56f35241523fBab1"
+const WETH: Address = "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2"
 
 export default function useCurvePreviewRedeem({
   asset,
@@ -34,30 +32,21 @@ export default function useCurvePreviewRedeem({
   const isArbitrumFamily = chainId === 313371 || chainId === 42161
   const poolCurveAddress = ARBI_CURVE_ADDRESS[asset] ?? "0x"
 
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let abi: any = undefined
-  if (poolCurveAddress === crvTriCryptoPoolAddress) {
-    abi = CurvePool3Assets
-  }
-  if (poolCurveAddress === crvTwoCryptoTokenAddress) {
-    abi = CurvePool2Assets
-  }
-  if (poolCurveAddress === fraxBpTokenAddress) {
-    abi = CurvePool2Assets
-  }
-
-  if (token === ETH && isArbitrumFamily) {
-    token = WETH_ARBI
-  }
-  if (token === ETH && !isArbitrumFamily) {
-    token = WETH
-  }
+  token =
+    token === ETH && isArbitrumFamily
+      ? WETH_ARBI
+      : token === ETH && !isArbitrumFamily
+      ? WETH
+      : token
 
   const underlyingAssets = useContractReads({
     contracts: [0, 1, 2, 3, 4].map((index) => ({
       address: poolCurveAddress,
       chainId,
-      abi,
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      abi: (poolCurveAddress === crvTriCryptoPoolAddress
+        ? CurvePool3Assets
+        : CurvePool2Assets) as any,
       functionName: "coins",
       args: [index],
     })),
@@ -90,4 +79,31 @@ export default function useCurvePreviewRedeem({
     data: parseInt(previewWithSlippage.toFixed(2)),
   }
   //}
+}
+
+function useCalcWithdrawOneCoin({
+  asset,
+  amount,
+  index,
+  enabled,
+}: {
+  asset: Address
+  amount: string
+  index: number
+  enabled: boolean
+}) {
+  const chainId = useActiveChainId()
+  const poolCurveAddress = ARBI_CURVE_ADDRESS[asset] ?? "0x"
+
+  return useContractRead({
+    chainId,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    abi: (poolCurveAddress === crvTriCryptoPoolAddress
+      ? CurvePool3Assets
+      : CurvePool2Assets) as any,
+    address: poolCurveAddress,
+    enabled: !!asset && enabled,
+    functionName: "calc_withdraw_one_coin",
+    args: [BigInt(amount), index],
+  })
 }
